+++ apps/server/package.json
@@ -14,7 +14,7 @@
     "dev:mock": "MOCK=true yarn dev",
     "watch": "yarn build --skip-initial-build --watch & nodemon build/index",
     "build": "babel src --out-dir build --extensions '.ts' --verbose",
-    "lint": "eslint src/**/*.ts",
+    "lint": "eslint \"src/**/*.ts\"",
     "typecheck": "tsc --noEmit",
     "test": "jest --passWithNoTests",
     "test:e2e": "concurrently 'yarn start:mock' 'wait-on tcp:3004 && yarn test:e2e:local' --kill-others --kill-others-on-fail --raw --success first",
