+++ apps/server/src/schema/album/index.ts
@@ -1,0 +1,2 @@
+export * from "./albumTypes";
+export * from "./albumQuery";
