+++ apps/server/src/schema/index.ts
@@ -1,5 +1,3 @@
-export * from "./photo";
 export * from "./album";
-export * from "./query";
 export * from "./auth";
 export * from "./user";
