+++ apps/server/src/schema/user/userResolvers.ts
@@ -1,6 +1,6 @@
-import { NexusGenFieldTypes } from "@howdypix/graphql-schema/schema";
+import { NexusGenFieldTypes } from "@howdypix/graphql-schema/schema.d";
 import { appDebug } from "@howdypix/utils";
-import { ApolloContext } from "../../types";
+import { ApolloContext } from "../../types.d";
 
 const debug = appDebug("gql:user");
